bill: fix test type expectancy

+++ bill/billt_test.go
@@ -129,5 +129,5 @@ func (s *S) TestUnbHoursBetweenStartAndEndWithEndOnSameDay(c *check.C) {
 func (s *S) TestUnbillableTimeBetweenSE(c *check.C) {
 	start := time.Date(2018, time.September, 4, 21, 0, 0, 0, time.UTC)
 	end := time.Date(2018, time.September, 5, 23, 0, 0, 0, time.UTC)
-	c.Assert(unbillableTimeBetweenSE(start, end), check.Equals, 480)
+	c.Assert(unbillableTimeBetweenSE(start, end), check.Equals, float64(480))
 }
